+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### Fixed
 - Fix bug in `Asset.from_model` where passed in asset ID was not used when creating a gwei or wei asset.
 
+### Added
+
+- Add `FundOperation` and `FundQuote` classes to support wallet funding.
+
 ## [0.10.3] - 2024-11-07
 
 ### Added